todo_server/requirements.txt
Django~=2.1.4
djangorestframework~=3.9.2
djangorestframework-simplejwt
django-cors-headers~=3.2.1
psycopg2-binary
